decouple extra functions into another header
